+++ src/widget.cpp
@@ -1498,7 +1498,7 @@ void NWidgetMatrix::SetClicked(int clicked)
 		int vpos = (this->clicked / this->widgets_x) * this->widget_h; // Vertical position of the top.
 		/* Need to scroll down -> Scroll to the bottom.
 		 * However, last entry has no 'this->pip_inter' underneath, and we must stay below this->sb->GetCount() */
-		if (this->sb->GetPosition() < vpos) vpos += this->widget_h - this->pip_inter - 1;
+		if ((int)this->sb->GetPosition() < vpos) vpos += this->widget_h - this->pip_inter - 1;
 		this->sb->ScrollTowards(vpos);
 	}
 }
